requirements.txt
aiohttp==3.7.4.post0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
asttokens==2.0.5
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.6.0
azure-common==1.1.28
azure-core==1.24.1
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
chardet==4.0.0
charset-normalizer==2.0.11
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==4.0.4
django-ckeditor==6.0.0
django-cors-headers==3.10.1
django-email-verification==0.1.0
django-ip-geolocation==1.6.1
django-js-asset==1.2.2
django-jsonfield==1.4.1
django-oauth-toolkit==1.7.0
django-phonenumber-field==6.0.0
django-phonenumbers==1.0.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
drf-social-oauth2==1.2.1
ecdsa==0.17.0
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
geoip2==4.1.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
imutils==0.5.4
incremental==21.3.0
ipykernel==6.13.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isodate==0.6.1
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.5.1
jupyter==1.0.0
jupyter-client==7.3.1
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
jwcrypto==1.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
maxminddb==2.0.3
mistune==0.8.4
msrest==0.6.21
multidict==5.1.0
mysqlclient==2.1.0
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.11
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
phonenumbers==8.12.40
pickleshare==0.7.5
Pillow==8.3.1
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.1
psycopg2==2.8.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.12.0
PyJWT==2.3.0
pyOpenSSL==20.0.1
pyotp==2.6.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-jose==3.3.0
python3-openid==3.2.0
pytz==2021.3
pyzmq==23.0.0
qtconsole==5.3.0
QtPy==2.1.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
Send2Trash==1.8.0
service-identity==21.1.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.2.0
terminado==0.15.0
tinycss2==1.1.1
tk==0.1.0
toml==0.10.2
tornado==6.1
traitlets==5.2.1.post0
Twisted==21.2.0
txaio==21.2.1
typing_extensions==4.2.0
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
widgetsnbextension==3.6.0
wrapt==1.13.3
yarl==1.6.3
zope.interface==5.4.0
